Solve random bug: Add track_tota_hits=true and verify size of the result
Solve a random bug that leads to the fetching of a wrong number of documents. In this version the number of documents returned by the search query is compared to the number returned by the count API and the program fails directly if the number is different.  
The param "track_tota_hits=true" is passed to the search as it seems to improve the probability to return the correct number, the size request also seems to have an influence (high probability the bug with a small size request).